Solves #619
Replaces miss on purpose with dynamic accuracy aaa naming like requested in #619

Kept Miss on purpose on some of the comments, with an addition of /dynamic accuracy aaa.  That way if your brains anything like mine you can search for either to find it.  Tear it out if you want, or I can.  Idk I'm not your dad.